modules/core/eks/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.61
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.50
    # via aws-cdk-lib
aws-cdk-lambda-layer-kubectl-v23==2.0.3
    # via -r requirements.in
aws-cdk-lib==2.51.0
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v23
    #   cdk-nag
boto3==1.26.81
    # via -r requirements.in
botocore==1.29.92
    # via
    #   boto3
    #   s3transfer
cattrs==22.1.0
    # via jsii
cdk-nag==2.12.29
    # via -r requirements.in
constructs==10.0.91
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
deepmerge==1.1.0
    # via -r requirements.in
exceptiongroup==1.1.1
    # via cattrs
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.74.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
packaging==23.0
    # via -r requirements.in
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
pyyaml==6.0
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lambda-layer-kubectl-v23
    #   aws-cdk-lib
    #   jsii
typing-extensions==4.2.0
    # via
    #   cattrs
    #   jsii
urllib3==1.26.15
    # via botocore
